+++ src/cc/arch/riscv64/ir_riscv64.c
@@ -159,7 +159,7 @@ Vector *collect_caller_save_regs(unsigned long living) {
   return saves;
 }
 
-static void push_caller_save_regs(Vector *saves, int total) {
+static void push_caller_save_regs(Vector *saves, size_t total) {
   int offset = total;
   offset += saves->len * TARGET_POINTER_SIZE;
   for (int i = 0; i < saves->len; ) {
@@ -172,7 +172,7 @@ static void push_caller_save_regs(Vector *saves, int total) {
   }
 }
 
-static void pop_caller_save_regs(Vector *saves, int offset) {
+static void pop_caller_save_regs(Vector *saves, size_t offset) {
   if (saves->len <= 0)
     return;
   for (int i = saves->len; i > 0; ) {
@@ -901,7 +901,7 @@ static void ei_pusharg(IR *ir) {
 }
 
 static void ei_call(IR *ir) {
-  int total = ir->call->stack_args_size;
+  size_t total = ir->call->stack_args_size;
   push_caller_save_regs(ir->call->caller_saves, total);
 
   if (ir->call->label != NULL) {
